One wiring file per ES module, named <module>_initialize.py
Carbon had two wiring files a letter apart: terrestrial_carbon_initialization.py
(GEP task trees) and terrestrial_carbon_initialize.py (the ES-shock graft). The
other three ES modules have only _initialize.py, so carbon was the odd one out and
the pair invited a wrong import.

Merged: add_terrestrial_carbon_tasks moves into the existing file, which takes the
_initialize name to match pollination, erosion and fisheries. All five existing
functions in it are unchanged.

Follow-on: four files import that module by name -- terrestrial_carbon_tasks.py and
the three run_terrestrial_carbon*.py scripts -- so their import lines change too.
Module name only, no behaviour change anywhere.
